requirements.txt
BeautifulSoup==3.2.1
Jinja2==2.7.1
Mako==0.7.0
Markdown==2.1.1
MarkupSafe==0.15
MySQL-python==1.2.3
Pillow==1.1
argparse==1.2.1
chardet==1.0.1
django==1.1.3

distribute==0.6.24dev-r0
django-contact-form==0.3.1
docutils==0.5
mod-python==3.3.1
openpyxl==1.6.1
pyPdf==1.13
python-dateutil==1.5
python-memcached==1.48
reportbug==4.12.6
reportlab==2.7
roman==1.4.0
simplejson==3.1.0
sqlparse==0.1.7
svg2rlg==0.3
wsgiref==0.1.2
